* 'master' of git://scm.dev.nokia.troll.no/qt/qt-qml-team: (536 commits)
  Adapt autotests to commit e2773a9
  Declarative: Use qDebug() for debugger status updates
  Declarative: Unify output of warning messages in debug server
  Added missing license header.
  Adjust pathes once again.
  Fixed a link
  Fixed a link
  Split behavior and wigglytext examples
  Rename files to show up in docs.
  Whitespace changes and added depends.
  Fix up examples.
  Fixed some linking issues.
  Removed unecessary files from folders.
  Fixes for the manifest and links.
  Changed some of the short descriptions for WebView examples.
  Removed unecessary duplicate files.
  Symbian-linux crosscompilation case changes
  Fix link to WebView examples.
  Fix link to ListView and WebView example pages.
  Fixed lists to be ordered lists and added curly brackets
  ...

+++ src/declarative/graphicsitems/qdeclarativeflipable.cpp
@@ -228,9 +228,9 @@ void QDeclarativeFlipablePrivate::updateSceneTransformFromParent()
         if (current == QDeclarativeFlipable::Back && back)
             setBackTransform();
         if (front)
-            front->setOpacity((current==QDeclarativeFlipable::Front)?1.:0.);
+            front->setOpacity((current==QDeclarativeFlipable::Front)?qreal(1.):qreal(0.));
         if (back)
-            back->setOpacity((current==QDeclarativeFlipable::Back)?1.:0.);
+            back->setOpacity((current==QDeclarativeFlipable::Back)?qreal(1.):qreal(0.));
         emit q->sideChanged();
     }
 }
